refactor: migrate early access API to use new error handling structure and update imports; remove deprecated error and clerk modules
